Introduce basic handling of std imports
Web3 and std interfaces are downloaded and installed on compilation.
The handling is currently very simple, we just replace the import string
with the std code.

We need to improve that and have some regex, because here it could fail
if there's one more blank space between  `import` and the package.

This will be done in a future commit.

Also notice we had to tweak a bit to read std files as we need to handle
both reading from sbt for tests and for a jar file in production.