Handle comments with imports
I added some types especially `ParsedImport` who keep tracks of multiple
information of the parsing.

Some of the issue before this commit was because we were trying to
comment the import line, by doing a find and replace first, which could
fail in multi edge cases.

Now we are only relying on what's provided by the parser.