Report Unknown import` at compile time
Resolves: #40

This implies quite an amount of changes (sorry for long commit).

During parsing, import are now only represented as a `ParsedImport`.

At compile time, we perform the actual compilation to extract the
import's AST.

The tricky part was to preserve the file URI of each import, so we could
return them as `SourceCodeState` error, rather than a workspace error.