Allow `SourceCode.compile` to have both workpace and source code errors
With the special case of imports, that are not handled by `ralphc`, we
need to compile them separately and provide their AST to the compiler.

Previously we could only have 1 workspace error or the successful code,
but now we could have both error, the compilation saying for example
that an interface is not found, as well as some imports errors, like if an
import is not found (which is what trigger the non found interface), we
want to show both error at the same time.