Fix import handler and allow new line in import names
Resolves: #43

While fixing the new lines things, I realized that the way we were
replacing the import statement in the source code was not working.
We were trying to be smart and preserve the name of the import as a
comment, but there are so many edge cases, it's better to comment out
the full parsed import.