Fix separator length on Windows
Windows is supposed to send `\r\n` for line separator.
But while testing `ralph-lsp` on vscode on Windows, I realized that only
`\n` was send to the server.

So we can't use `System.lineSeparator` and we can't either only use `\n`
because we can't exclude that another client used `\r\n`