Fix inconsistent warnings
This is not the perfect solution, as we need to deep copy the AST, too
keep track of the original AST.

Also we still need to mutate just a bit to have access to all feature
like some go-to-definiton, I haven't debug that last part yet. Could be
improved in a later commit.

I tried to keep the fix in a separate file so it's easy to revert it.